requirements.txt
ipwhois
requests
python-whois
folium